Improve minBits calculation
More correctly account for the signed nature of the timestamp DoD.

A DoD of 3 cannot actually be encoded in two bits as 0b11, that would
be interpreted as the lowest possible value, -1. Thus, 3 would require
3 bits, 0b011.

A DoD of 4 _can_ be encoded in 3 bits 0b100 - this is contrary to
standard two's complement, in which it would normally encode -4.

This reflects the actual behaviour of Prometheus:
 * https://github.com/prometheus/prometheus/blob/release-2.26/tsdb/chunkenc/xor.go#L375
 * https://github.com/prometheus/prometheus/blob/release-2.26/tsdb/chunkenc/xor.go#L203