Provide a cross-index per-series sample count
Implemented as

 CrossIndexSampleIterator.getNumSamples()

or in Python

 len(samples)

e.g.,

 for series, samples in load("..."):
    print(len(samples))

Which calculates the number of samples in that series, across all
present block dirs/indexes without having to read all the actual sample
data.

This is possible because the chunks contain the number of samples,
which can simply be summed across all chunks in all block dirs for the
given series.