Manage Series through an aliased shared_ptr
This is done to avoid a SeriesSampleIterator outliving an Index, and
referencing invalid data.

This would be unlikely to occur in C++, as users would reasonably
expect iterators and references to be invalidated if the containing
datastructure is destroyed, but in contrast it poses a plausible
footgun in python if keep_alives can't be leveraged quite as desired.