Fix regex and Python function filters
Previously, Python callbacks were converted to a string as if by repr
of the form:
 <lambda whatever at 0xdeadbeef>

This was then used to exact match a label value - clearly this is
incorrect.

This patch additionally ensures callbacks created in C++, say through

 pyp.regex(...)

which calls

 Filter pdu::filter::regex(...)

are handled distinctly from Python callbacks. It is more efficient to
directly call the std::function Filter rather than calling up and down
through pybind as if it is any old python function.