Downgrade Boost to 1.74
The conan pre-built package for boost 1.76 requires statx, and statx is
not available in the manylinux2014 environment, making a widely
compatible Python wheel difficult.

If boost were configured and built in this env, 1.76 would probably be
fine, as statx would not be detected during configuration (or could be
disabled with BOOST_FILESYSTEM_DISABLE_STATX). However, this would
significantly increase build times in Github Actions.

Similarly, 1.77 adds a runtime statx check which may allow the use of
that version, but it is not available from conan center yet.