Initialise tsPrefix
tsPrefix was not always initialised as 0, so could contain garbage,
leading to an invalid tsPrefix. This lead to tsBitCount remaining
defaulted to 0, and a lot of subtle bugs.

This did not present when compiled with clang, suggesting clang may
initialise uint8_t to 0, and reinforces the need for test suites and CV
on multiple platforms.