Add buffer_protocol support for Series.samples
Renames Series.samples to Series.sample_iterator, which continues to
support lazy loading.

Adds a new Series.samples which eagerly reads all samples for a series,
and exposes the Python Buffer Protocol[1]. This allows simple interop
with numpy and other similar C extensions.

 import numpy

 data = pypdu.load(...)
 for series in data:
     samples_array = numpy.array(series.samples, copy=False)
     samples_mv = memoryview(series.samples)
     raw_bytes = samples_mv.tobytes()

As the buffer contains struct elements (Samples) a lot of memoryview
functionality is unfortunately not available.

[1]:https://docs.python.org/3/c-api/buffer.html