Improve irate handling of reset of monotonic counter value
If a counter is reset (e.g., a service restarts), the time series will
appear to drop. As counters are expected to be monotonic, a drop can be
assumed to indicate the value has returned to zero since the last
sample.

To optionally handle this in the same was as Prometheus, adds an
optional monotonic param:

 irate(expression, monotonic=False)

Using this avoids large negative rates.