Make numpy an optional runtime dependency
Samples can be exposed as a numpy array, but only if numpy is available
at runtime.

Previously, pypdu had a mandatory dependency on numpy, but this is
unnecessarily restrictive.

This patch makes pypdu use numpy if installed, but makes it an optional
dependency. This means pypdu can still be used on platforms where numpy
is not available (e.g., Python2 on up to date M1 macs, apparently numpy
is not simple to build there).