Refactor SeriesHandle to make SeriesSource available
Prevously, SeriesHandle only needed to expose a sample iterator,
as that was all the the existing use cases required.

Going forward, direct access to the underlying chunk data will
be required; to facilitate this, have SeriesHandle keep a ptr to the
SeriesSource. With this, ChunkRefs from the Series can be resolved to
actual chunks in the source's chunk cache.