Assume malformed head chunk is partially written data
Head chunks and the WAL may not be completely flushed to disk if
e.g., Prometheus crashes.

Partial records may exist at the end of the file; allow this and assume
there is not further valid data.

A future change should validate CRCs also, and stop reading on the first
incorrect CRC.